Consolidate to iterative overlay root
