+++ src/storage.rs
@@ -1,7 +1,6 @@
 pub mod debug;
 pub mod engine;
 pub mod overlay_root;
-pub mod overlay_root_iterative;
 pub mod proofs;
 mod test_utils;
 pub mod value;
